fix: синхронизированы типы моделей со схемой PostgreSQL
- int64 → int32 для всех ID во всех моделях и relation-типах
- Tournament: добавлено поле CountryID, Month: int8 → int16
- database/init.sql заменен на актуальный pg_dump

+++ internal/domain/models/city.go
@@ -1,7 +1,7 @@
 package models
 
 type City struct {
-	ID          int64  `db:"id"`
+	ID          int32  `db:"id"`
 	Name        string `db:"name"`
 	Description string `db:"description"`
 	RepublicID  *int64 `db:"republic_id"`
