+++ internal/domain/models/sportclub.go
@@ -1,11 +1,11 @@
 package models
 
 type SportClub struct {
-	ID             int64  `db:"id"`
+	ID             int32  `db:"id"`
 	Name           string `db:"name"`
 	FullName       string `db:"full_name"`
 	Founded        string `db:"founded"`
-	CityID         int64  `db:"city_id"`
+	CityID         int32  `db:"city_id"`
 	Region         string `db:"region"`
 	HeadCoach      string `db:"head_coach"`
 	Description    string `db:"description"`
